refactor(phase3): extract config utilities from bin/cli.js
Part of #89 - Phase 3: Extract config utilities

Extracted config utility functions:
- loadFingerprint() + applyFingerprintToConfig() → lib/utils/fingerprint.js (32 lines)
- loadProjectConfigFile() + writeProjectConfigFile() → added to lib/utils/project-config.js
- deepMerge() → already in project-config.js, now exported

Progress:
- bin/cli.js: 528 → 473 lines (-55 lines)
- Total progress: 769 → 473 lines (-296 lines, -38%)
- All 554 tests passing

Third of 6 PRs to refactor bin/cli.js to a thin router (~40 lines)
