+++ lib/utils/project-config.js
@@ -123,7 +123,24 @@ function readProjectConfig(context) {
   return merged;
 }
 
+function loadProjectConfigFile(cwd) {
+  const file = path.join(cwd, '.ai-coding-guide.json');
+  try {
+    const raw = fs.readFileSync(file, 'utf8');
+    return { file, json: JSON.parse(raw) };
+  } catch {
+    return { file, json: DEFAULTS };
+  }
+}
+
+function writeProjectConfigFile(file, json) {
+  fs.writeFileSync(file, JSON.stringify(json, null, 2) + '\n');
+}
+
 module.exports = {
   readProjectConfig,
-  DEFAULTS
+  DEFAULTS,
+  deepMerge,
+  loadProjectConfigFile,
+  writeProjectConfigFile
 };
